move hex encrypt to general util
