+++ cli/src/commands.rs
@@ -24,6 +24,7 @@ use codec::{Decode, Encode};
 use ita_stf::ShardIdentifier;
 use itc_rpc_client::direct_client::DirectApi;
 use itp_node_api_extensions::{PalletTeerexApi, TEEREX};
+use itp_utils::encrypt_to_hex_bytes;
 use log::*;
 use my_node_runtime::{Balance, BalancesCall, Call, Event, Hash};
 use sp_application_crypto::{ed25519, sr25519};
