+++ core-primitives/utils/src/error.rs
@@ -27,6 +27,8 @@ pub type Result<T> = core::result::Result<T, Error>;
 pub enum Error {
 	#[error("Insufficient buffer size. Actual: {0}, required: {1}")]
 	InsufficientBufferSize(usize, usize),
+	#[error("Encryption faild due to: {0}")]
+	Encryption(sgx_types::sgx_status_t),
 	#[error(transparent)]
 	Other(#[from] Box<dyn std::error::Error + Sync + Send + 'static>),
 }
